# docs/openapi-rbac.yaml
openapi: 3.0.0
info:
  title: ATA CRM - RBAC API
  version: 1.0.0
  description: Role-Based Access Control API endpoints

servers:
  - url: http://localhost:3005/api
    description: Development server

security:
  - bearerAuth: []
  - cookieAuth: []

paths:
  /auth/me:
    get:
      summary: Get current user with permissions
      tags:
        - Authentication
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: User information with permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
                      role:
                        type: string
                      companyId:
                        type: integer
                      permissions:
                        type: array
                        items:
                          type: string
                      roles:
                        type: array
                        items:
                          type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /rbac/roles:
    get:
      summary: Get all roles
      tags:
        - RBAC - Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: includeSystem
          in: query
          schema:
            type: boolean
          description: Include system roles
        - name: companyOnly
          in: query
          schema:
            type: boolean
          description: Show only company roles
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new role
      tags:
        - RBAC - Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /rbac/roles/{id}:
    patch:
      summary: Update a role
      tags:
        - RBAC - Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a role
      tags:
        - RBAC - Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Role deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /rbac/permissions:
    get:
      summary: Get all permissions
      tags:
        - RBAC - Permissions
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: resource
          in: query
          schema:
            type: string
          description: Filter by resource
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new permission
      tags:
        - RBAC - Permissions
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionRequest'
      responses:
        '201':
          description: Permission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /rbac/users/{userId}/roles:
    get:
      summary: Get user roles
      tags:
        - RBAC - User Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserRole'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Assign role to user
      tags:
        - RBAC - User Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '201':
          description: Role assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Remove role from user
      tags:
        - RBAC - User Roles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: roleId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Role removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /rbac/audit-logs:
    get:
      summary: Get audit logs
      tags:
        - RBAC - Audit Logs
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: action
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token

  schemas:
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        isSystem:
          type: boolean
        companyId:
          type: integer
          nullable: true
        permissions:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        userCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateRoleRequest:
      type: object
      required:
        - name
        - displayName
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        companyId:
          type: integer
        permissionIds:
          type: array
          items:
            type: integer

    UpdateRoleRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        permissionIds:
          type: array
          items:
            type: integer

    RoleResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Role'
        message:
          type: string

    Permission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        category:
          type: string
        resource:
          type: string
        action:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreatePermissionRequest:
      type: object
      required:
        - name
        - displayName
        - category
        - resource
        - action
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        category:
          type: string
          minLength: 1
          maxLength: 50
        resource:
          type: string
          minLength: 1
          maxLength: 50
        action:
          type: string
          minLength: 1
          maxLength: 50

    PermissionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Permission'
        message:
          type: string

    UserRole:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        isDefault:
          type: boolean
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: integer
        isActive:
          type: boolean

    AssignRoleRequest:
      type: object
      required:
        - roleId
      properties:
        roleId:
          type: integer
        expiresAt:
          type: string
          format: date-time

    UserRoleResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/UserRole'
        message:
          type: string

    AuditLogResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            logs:
              type: array
              items:
                type: object
            pagination:
              type: object

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Authentication required"

    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Insufficient permissions. Required: role.manage"

    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
              details:
                type: array

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Role not found"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Failed to process request"